backend/requirements.txt
fastapi
python-multipart
uvicorn
regex
httpx
pillow
werkzeug